+++ web/src/pages/Restore.svelte
@@ -30,7 +30,7 @@
   onMount(() => {
     loadJobs()
     const unsub = onWsMessage((msg) => {
-      if (msg.type === 'job_run_completed') {
+      if (msg.type === 'job_run_completed' || msg.type === 'import_completed') {
         loadJobs()
       }
     })